requirements.txt
dj-database-url==0.4.2
Django==1.11.3
djangorestframework==3.6.3
gunicorn==19.7.1
mysqlclient==1.3.10
pytz==2017.2
whitenoise==3.3.0
